+++ refdrmaster.py
@@ -15,6 +15,7 @@ import os
 from editrefdr import Ui_editrefdrForm
 from PyQt5.QtCore import QTime, QTimer
 from PyQt5.QtWidgets import QTableWidget,QTableWidgetItem
+from PyQt5.QtWidgets import QMessageBox
 
 class Ui_refdrmasterForm(object):
     def setupUi(self, Form):
@@ -347,11 +348,21 @@ class Ui_refdrmasterForm(object):
 
 
     def delete_refdr(self, doc_code):
-       
-       # Connect to the database
+      confirm_dialog = QMessageBox()
+      confirm_dialog.setIcon(QMessageBox.Question)
+      confirm_dialog.setText("Are you sure you want to delete this refdr?")
+      confirm_dialog.setWindowTitle("Confirm Deletion")
+      confirm_dialog.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+  
+      # Show the dialog and wait for the user's response
+      response = confirm_dialog.exec_()
+  
+      # If the user confirms deletion, proceed with deletion
+      if response == QMessageBox.Yes:
+        # Connect to the database
        conn = sqlite3.connect('patient_data.db')
        cursor = conn.cursor()
-
+       
        # Delete patient data from the database
        cursor.execute("DELETE FROM refdr WHERE DoctorCode = ?", (doc_code,))
        conn.commit()
